Layout of collectionViewCells
Was trying to get the positioning of the cells just right, probably still need to do some other styling and being able to visualize city data at least mock data until we get our api.  Also on the todo, i need to make a way to be able to save cities from the map and save these to an array so i can display the users actual favorite places vs the mock data i have now.